+++ VoxelWorld/source/2d/Rect.cpp
@@ -13,7 +13,7 @@ GLuint Rect::vao = 0;
 GLuint Rect::vbo = 0;
 std::shared_ptr<ShaderProgram> Rect::shaderProgram = std::shared_ptr<ShaderProgram>();
 
-Rect::Rect(int x, int y, int width, int height) : x(x), y(y), width(width), height(height) {
+Rect::Rect(float x, float y, float width, float height) : x(x), y(y), width(width), height(height) {
     if (!Rect::shaderProgram) {
         Rect::init();
     }
@@ -57,8 +57,8 @@ void Rect::render(glm::mat4 projectionMatrix, bool wireframe) {
     Rect::shaderProgram->use();
 
     glm::mat4 modelMatrix = glm::mat4(1.0f);
-    modelMatrix = glm::translate(modelMatrix, glm::vec3(static_cast<float>(x), static_cast<float>(y), 0.0f));
-    modelMatrix = glm::scale(modelMatrix, glm::vec3(static_cast<float>(width), static_cast<float>(height), 1.0f));
+    modelMatrix = glm::translate(modelMatrix, glm::vec3(x, y, 0.0f));
+    modelMatrix = glm::scale(modelMatrix, glm::vec3(width, height, 1.0f));
 
     auto mvp = projectionMatrix * modelMatrix;
